requirements.txt
torch
flask
numpy
opencv-python
pillow